# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top All Money API",
        "description": "Documentation de l'API top all money",
        "contact": {
            "email": "support@tonapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        },
        {
            "url": "https://api.tonapp.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isDirty",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "updatedAt:desc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 200,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a product",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show a product",
                "operationId": "c87fb2e4d617e7b36c712cad358acdb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "416e4959a3919547eb609d87e3db9135",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Soft delete",
                "operationId": "b49ce9f3cb168b233245548fb4ef36d3",
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Partially update a product",
                "operationId": "7b6a99d3d6d5c75a1ef063b213e8d068",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/products/bulk": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Bulk upsert products",
                "operationId": "5b8aa874fb98986ca8686da2d995d25f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ProductUpdateRequest"
                                                },
                                                {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid",
                                                            "nullable": true
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "enum": [
                                                                "CREATE",
                                                                "UPDATE",
                                                                "DELETE"
                                                            ]
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/products/import": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Import products from CSV",
                "operationId": "dd7d5171cdd91a805518de03df641cc4",
                "responses": {
                    "200": {
                        "description": "Imported"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductCreateRequest": {
                "required": [
                    "code",
                    "name"
                ],
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "barcode": {
                        "type": "string",
                        "nullable": true
                    },
                    "unitId": {
                        "type": "string",
                        "nullable": true
                    },
                    "categoryId": {
                        "type": "string",
                        "nullable": true
                    },
                    "defaultPrice": {
                        "type": "integer",
                        "example": 0
                    },
                    "purchasePrice": {
                        "type": "integer",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "ProductUpdateRequest": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "barcode": {
                        "type": "string",
                        "nullable": true
                    },
                    "unitId": {
                        "type": "string",
                        "nullable": true
                    },
                    "categoryId": {
                        "type": "string",
                        "nullable": true
                    },
                    "defaultPrice": {
                        "type": "integer"
                    },
                    "purchasePrice": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}